+++ lib/src/utils/utils.dart
@@ -1,4 +1,5 @@
 export 'assets_loader.dart';
+export 'context_utils.dart';
 export 'interfases.dart';
 export 'list_utils.dart';
 export 'math_utils.dart';